+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Designer UI improvements
 	- Field names and types are now truncated when they are too long
 	- Limit tables to displaying 7 fields, with a button to view more
+- Add button to load query from a file
 - Replaced the window pinning toolbar button with a settings entry and hotkey (F11)
 - Fixed web app always starting in light theme for a split second